:green_heart: Remove redundant type cast

+++ src/InboundClient.php
@@ -158,7 +158,7 @@ class InboundClient extends EventEmitter
                     'error',
                     [new ReactESLException(
                         'Unexpected reply received (ENOMSG) ' . $contentType,
-                        defined('SOCKET_ENOMSG') ? (int)SOCKET_ENOMSG : 42
+                        defined('SOCKET_ENOMSG') ? SOCKET_ENOMSG : 42
                     )]
                 );
             } else {
